made contract controller api compliant

+++ upromisapi/Models/ContractRepository.cs
@@ -17,19 +17,22 @@ namespace upromiscontractapi.Models
             _context = context;
 
         }
-        public IQueryable<ContractDTO> List => _context.Contracts.Select(c => Transformers.Transform(c, new ContractDTO() { Modifier = "Unchanged"} ) ).AsQueryable();
+        public IQueryable<ContractDTO> List => _context.Contracts.Select(c => Transformers.Transform(c, new ContractDTO() { Modifier = "Unchanged" })).AsQueryable();
 
-        public async Task<APIResult<ContractDTO>> Post(SaveMessage<ContractDTO> rec )
+        public async Task<ContractDTO> Post(SaveMessage<ContractDTO> rec)
         {
             // TODO: add validations
             // TODO: check for new users to create in the team composition list
 
             Contract contract = Transformers.Transform(rec.DataSubject, new Contract()) as Contract;
-            
-            contract.PaymentInfo.ForEach(pi => { pi.Contract = contract;
+
+            contract.PaymentInfo.ForEach(pi =>
+            {
+                pi.Contract = contract;
             });
 
-            contract.TeamComposition.ForEach(pi => {
+            contract.TeamComposition.ForEach(pi =>
+            {
                 pi.Contract = contract;
             });
 
@@ -38,38 +41,48 @@ namespace upromiscontractapi.Models
             contract.ParentContract = _context.Contracts.Find(rec.DataSubject.ParentContractID);
 
             _context.Contracts.Add(contract);
-            
+
             await _context.SaveChangesAsync();
 
-            return new APIResult<ContractDTO>() { ID = rec.DataSubject.ID, Success = true, DataSubject = Transformers.Transform(contract, new ContractDTO()), Message = "Post was performed" };
+            return Transformers.Transform(contract, new ContractDTO());
         }
 
-        public async Task<APIResult<ContractDTO>> Put(SaveMessage<ContractDTO> rec )
+        public async Task<ContractDTO> Put(SaveMessage<ContractDTO> rec)
         {
             var ctr = await _context.Contracts.Where(c => c.ID == rec.ID)
                 .Include("PaymentInfo")
                 .Include("TeamComposition")
                 .FirstOrDefaultAsync();
 
+            if (ctr == null)
+            {
+                return null;
+            }
+
             ctr = Transformers.Transform(rec.DataSubject, ctr);
 
             await _context.SaveChangesAsync();
 
-            return new APIResult<ContractDTO>() { ID = ctr.ID, DataSubject=Transformers.Transform(ctr, new ContractDTO()), Success = true, Message = "Put was performed" };
+            return Transformers.Transform(ctr, new ContractDTO());
         }
 
-        public async Task<APIResult<ContractDTO>> Delete(SaveMessage<ContractDTO> rec )
+        public async Task<bool> Delete(SaveMessage<ContractDTO> rec)
         {
             var ctr = _context.Contracts.FirstOrDefault(c => c.ID == rec.ID);
+
+            if (ctr == null)
+            {
+                return false;
+            }
 
             _context.Contracts.Remove(ctr);
 
             await _context.SaveChangesAsync();
 
-            return new APIResult<ContractDTO>() { ID = rec.ID, Success = true, DataSubject = null, Message = "Delete was performed" };
+            return true;
         }
 
-        public async Task<APIResult<ContractDTO>> Get(int id)
+        public async Task<ContractDTO> Get(int id)
         {
             var contract = await _context.Contracts.Where(c => c.ID == id)
                 .Include("PaymentInfo")
@@ -78,7 +91,7 @@ namespace upromiscontractapi.Models
 
             var ctr = Transformers.Transform(contract, new ContractDTO() { Modifier = "Unchanged" });
 
-            return new APIResult<ContractDTO>() { ID = id, Success = true, DataSubject = ctr, Message = "Get was performed" };
+            return ctr;
         }
     }
 }
